Changed ProcessBuilder to use the Java process's I/O
wget was erroring out, but it wasn't being surfaced in Java. I finally traced it back to the StreamGobbler and felt it was much simpler just to have the wget process be able to direct its I/O to the existing Java's I/O, so I'd directly see if I'd screwed up the command-line arguments again.