Removed the rel_path variable
I couldn't find a way to make wget save all the image/css/js assets to a separate directory. Best thing I could do was have it dump everything, including the html page, into a single directory. I'm not sure if this will break on some sites, but it seemed to work fine on the sites I tested.