Reverted to using the hashed URL as part of the filename
Also found a way to tell wget how to specify the default page name (--default-page=name), since it was automatically setting a name based on the URL, which made it difficult to deal with programmatically.

Also, I realized the URL hash is a decent idea for later usage, but I still wanted the page title included for human parsing.